Refactor module placements
