+++ modules/monoxide.rb
@@ -1,8 +1,5 @@
 # frozen_string_literal: true
 
-require_relative 'statistics'
-
-# Monoxide
 class Monoxide < SensorModule
   def initialize(name, baseline)
     super(name, baseline)
