+++ utils/statistics.rb
